Move CPU & mem into their own modules
