+++ termux_hardware_stats/stats/termux_cpu.py
@@ -1,3 +1,34 @@
+from collections import namedtuple
+from itertools import takewhile
+from dataclasses import dataclass
+
+FREQ_KEYS = ['curr_freq', 'min_freq', 'max_freq']
+
+FPATHS = {
+    'global_state': '/sys/devices/system/cpu/cpu{:d}/core_ctl/global_state',
+	'curr_freq': '/sys/devices/system/cpu/cpu{:d}/cpufreq/scaling_cur_freq',
+	'min_freq': '/sys/devices/system/cpu/cpu{:d}/cpufreq/scaling_min_freq',
+	'max_freq': '/sys/devices/system/cpu/cpu{:d}/cpufreq/scaling_max_freq',
+    'mem_info': '/proc/meminfo',
+}
+
+GlobalState = namedtuple(
+    'GlobalState', [
+        'CPU',
+        'Online',
+        'Isolated',
+        'FirstCPU',
+        'BusyPercentage',
+        'IsBusy',
+        'NotPreferred',
+        'NrRunning',
+        'ActiveCPUs',
+        'NeedCPUs',
+        'NrIsolatedCPUs',
+        'Boost',
+        'OPBoost',
+    ]
+)
 CPUFrequency = namedtuple(
     'CPUFrequency', [
         'current',
@@ -5,6 +36,7 @@ CPUFrequency = namedtuple(
         'max',
     ]
 )
+
 
 @dataclass
 class CPUGlobalStateReader:
@@ -18,8 +50,6 @@ class CPUGlobalStateReader:
                 stats_chunk = list(takewhile(lambda x: not x.startswith('CPU'), istream))
                 yield GlobalState(*list(map(lambda x: x.strip().split(': ')[1], stats_chunk)))._asdict()
 
-FREQ_KEYS = ['curr_freq', 'min_freq', 'max_freq']
-
 @dataclass
 class CPUFrequencyReader:
     cpu_count: int
@@ -27,6 +57,6 @@ class CPUFrequencyReader:
     def load_for_core(n):
         return [int(next(open(FPATHS[k].format(n)))) for k in FREQ_KEYS]
 
-    def load_all():
-        for i in range(TermuxHardwareStats.cpu_count()):
+    def load_all(self):
+        for i in range(self.cpu_count):
             yield CPUFrequency(*CPUFrequencyReader.load_for_core(i))._asdict()
